requirements.txt
amqp==5.2.0
asgiref==3.7.2
billiard==3.6.4.0
boto3==1.26.50
botocore==1.29.165
celery==5.2.7
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
crispy-bootstrap5==2023.10
Django==4.1.4
django-appconf==1.0.6
django-crispy-forms==2.1
django-imagekit==5.0.0
django-storages==1.13.2
dnspython==2.4.2
idna==3.6
jmespath==1.0.1
kombu==5.3.4
numpy==1.26.2
opencv-contrib-python==4.7.0.68
pilkit==3.0
Pillow==9.4.0
prompt-toolkit==3.0.41
pymongo==4.3.3
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
s3transfer==0.6.2
six==1.16.0
sketchify==0.0.8
sqlparse==0.4.4
stripe==5.0.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==1.26.18
vine==5.1.0
wcwidth==0.2.12
